+++ src/main/java/uk/ac/nott/cs/comp2013/froggerApp/model/gameObjects/actors/factories/WetTurtleFactory.java
@@ -5,7 +5,7 @@ import uk.ac.nott.cs.comp2013.froggerApp.model.GameConfig.*;
 import uk.ac.nott.cs.comp2013.froggerApp.model.gameObjects.actors.level.WetTurtle;
 
 public class WetTurtleFactory {
-    public static WetTurtle createWetTurtle(int size, int xpos, int row, double speed) {
+    public static WetTurtle createWetTurtle(int size, double xpos, int row, double speed) {
         WetTurtle wetTurtle = new WetTurtle();
         wetTurtle.setTurtleImages(
                 new Image(TurtleConfig.IMAGE_PATHS.get("animation1"), size, size,true,true),
